requirements.txt
boto3~=1.34.134
google-cloud-storage
configparser~=7.0.0
botocore~=1.34.134